requirements.txt
mysql-connector==2.2.9
mysql-connector-python==9.4.0
pandas==2.3.0
pillow==11.3.0
plotly==6.2.0
streamlit==1.47.1
toml==0.10.2

